refactor: Remove handling of radio-buttons state from <Question />
Changed the handling of the state of the radio buttons in the <Question /> component
to the higher-level <Game /> component.
This was done so as to be able to handle for changes in the state when game is over.
When the <Question /> held the state, on game over it rerendered the component,
thus initializing the state to default therefore there were no correct answers
as no option across the game is selected by default.