feat: Change the game mode from 'game-over' to 'continue game' when the user selects options
Changing the functionality set on the previous commit.
When a game play is on its status is 0
When its game over its 1
But if the user decided to play again after game is over, a different status (0.5)
is used. This is to enable the timer that will be implemented display correctly on
each mode