feat: Implement the whole app reducer, context and localStorage
Basically rewrite the whole app using a central reducer to manage state
and context for providing the state to all the component 'pages'. Also implemented adding the state to storage so the changes made are not lost on refresh.

The changes made in this commit doesn't follow standard of keeping commits specific and
centered on one issue. For this I apologize.